websearch/requirements.txt
textual>=0.58.0
httpx>=0.27.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
